test: 단위테스트를 통한 JPA 기반의 KeyValueRepository 검증
- 단위 테스트 환경에서는 H2 데이터베이스를 사용합니다.
- 테스트를 위한 의존성을 추가했습니다.
- 사용되지 않는 기존의 테스트 코드는 제거했습니다.
